feat: enable helm deps (#54)
Enables helm dependencies for kustomize.  Some of the client was redone to support edge cases that came up.  Primarily supporting default namespaces (will be updated later to isolate to specific namespaces), and some renaming.  There's still an issue in the apply step that needs to be addressed as it appears that when crds are loaded the api resources are not reflecting (reloading) the additional crds.  It will fail the first time with a not found (referencing the kind) but then install if it is run again.