feat: add env configuration command (#69)
Adds a new config command for managing environment configs.   There is a new assumption that environments are a global resource to be shared across projects.  An additional assumption is that the controller should never need a default environment and should fail a sync if the the config for a specific environment does not exist.

* [x] Add a config command that allows users to create the config from seactl.
* [x] Rename `SeawayConfig` to `EnvironmentConfig` to more accurately reflect it's purpose.
* [x] Add the registry to the installer (I'd forgotten to do that on the last PR and it's needed to test out the config with sync)

Chores:
* [x] Move the tailscale and registry kustomization folders into seaway folder so they can be used with the install generator.