ctx/fix(sync): address error on force (#107)
Fixes an issue where the sync would fail on force after the environment was deleted.  The error would look similar to this:

```shell
◉ Uploading archive
  ◦ Size: 3082
  ◦ Revision: 60014ebda6c4f17fcc636398126e109f-1
◉ Removing existing environment
◉ Deploying
✘ error modifying environment: Operation cannot be fulfilled on environments.seaway.ctx.sh "seaway-example": StorageError: invalid object, Code: 4, Key: /registry/seaway.ctx.sh/environments/default/seaway-example, ResourceVersion: 0, AdditionalErrorMsg: Precondition failed: UID in precondition: 499d0e51-f816-4567-afb7-97a6341bf28c, UID in object meta:
```
The cause was that I was reusing the object which had been modified from the delete command when calling create which included fields that were not expected to be set.  This ensures that a new object is used for the creation.